Convert simpletest to phpunit
Add composer and gulp to run phpunit

Changes to src files:
-Require HgRexceptions.php in BundleHelper.php
-Change ValidationException in HgRunner::getRevisionsInternal from "quantity parameter much be larger than 0" to "quantity parameter must be larger than 0"
