+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "hgresume",
+  "version": "3.0.0",
+  "description": "Provides a server API to transfer Hg bundles that can be retried/retransmitted",
+  "main": "api/index.php",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/sillsdev/hgresume.git"
+  },
+  "author": "",
+  "license": "MIT",
+  "devDependencies": {
+    "gulp": "github:gulpjs/gulp#4.0",
+    "gulp-concat": "~2.4.1",
+    "gulp-util": "~2.2.14",
+    "gulp-phpunit": "^0.21.4",
+    "async": "~0.9.0",
+    "lodash.template": "^2.4.1",
+    "yargs": "^6.3.0"
+  }
+}
