+++ test/phpunit.xml
@@ -1,0 +1,46 @@
+<phpunit
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/4.5/phpunit.xsd"
+    backupGlobals="true"
+    backupStaticAttributes="false"
+    bootstrap="testconfig.php"
+    cacheTokens="false"
+    colors="auto"
+    convertErrorsToExceptions="true"
+    convertNoticesToExceptions="true"
+    convertWarningsToExceptions="true"
+    forceCoversAnnotation="false"
+    mapTestClassNameToCoveredClassName="false"
+    processIsolation="false"
+    stopOnError="false"
+    stopOnFailure="false"
+    stopOnIncomplete="false"
+    stopOnSkipped="false"
+    stopOnRisky="false"
+    timeoutForSmallTests="1"
+    timeoutForMediumTests="10"
+    timeoutForLargeTests="60"
+    verbose="true"
+>
+    <php>
+        <ini name="xdebug.show_exception_trace" value="0"/>
+    </php>
+    <testsuites>
+        <testsuite name="hgresume Test Suite">
+            <directory>./</directory>
+        </testsuite>
+    </testsuites>
+    <groups>
+        <exclude>
+            <group>explicit</group>
+        </exclude>
+    </groups>
+    <filter>
+        <whitelist processUncoveredFilesFromWhitelist="true">
+            <directory suffix=".php">../src</directory>
+            <exclude>
+                <directory suffix=".php">../src/vendor</directory>
+            </exclude>
+        </whitelist>
+    </filter>
+</phpunit>
